added FullCalendar dependencies
